WE-58: add a layer in WorkOrderInvoiceDetailListener to properly use Transactional annotation
